Move ReactNativeEngine to its own file and clean it up more
